Revert "Workout list"
