+++ android/app/src/main/AndroidManifest.xml
@@ -1,20 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.flutter_application_2">
-    <uses-permission android:name="android.permission.QUERY_ALL_PACKAGES" />
-    <queries>
-    <intent>
-        <action android:name="android.intent.action.VIEW" />
-        <data android:scheme="https" />
-    </intent>
-    <intent>
-        <action android:name="android.intent.action.DIAL" />
-        <data android:scheme="tel" />
-    </intent>
-    <intent>
-        <action android:name="android.intent.action.SEND" />
-        <data android:mimeType="*/*" />
-    </intent>
-    </queries>
    <application
         android:label="flutter_application_2"
         android:name="${applicationName}"
