+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,11 +6,9 @@ import FlutterMacOS
 import Foundation
 
 import path_provider_macos
-import url_launcher_macos
 import sqflite
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
   SqflitePlugin.register(with: registry.registrar(forPlugin: "SqflitePlugin"))
-  UrlLauncherPlugin.register(with: registry.registrar(forPlugin: "UrlLauncherPlugin"))
 }
