+++ pubspec.yaml
@@ -43,9 +43,6 @@ dependencies:
   get: ^4.6.5
   accordion: ^2.5.1
   google_fonts: ^3.0.1
-  localstorage: ^4.0.0+1
-  numberpicker: ^2.1.1
-  url_launcher: ^6.1.7
   path_provider: ^2.0.11
   sqflite: ^2.2.2
